refined memento into a single module of it's own
